refactor objective class and decoupled class files for halfGraph/model trussEnv/moo
